create base package and move the base classes
