+++ app/src/main/java/com/sanaa/tudee_assistant/presentation/composable/bottomSheet/task/taskDetailsBottomSheet/TaskDetailsBottomSheetViewModel.kt
@@ -4,10 +4,10 @@ import com.sanaa.tudee_assistant.domain.model.Task
 import com.sanaa.tudee_assistant.domain.service.CategoryService
 import com.sanaa.tudee_assistant.domain.service.StringProvider
 import com.sanaa.tudee_assistant.domain.service.TaskService
+import com.sanaa.tudee_assistant.presentation.base.BaseViewModel
 import com.sanaa.tudee_assistant.presentation.model.TaskUiStatus
 import com.sanaa.tudee_assistant.presentation.model.mapper.toDetailsState
 import com.sanaa.tudee_assistant.presentation.model.mapper.toTask
-import com.sanaa.tudee_assistant.presentation.utils.BaseViewModel
 import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.collectLatest
