+++ app/src/main/java/com/sanaa/tudee_assistant/presentation/screen/main/MainViewModel.kt
@@ -1,7 +1,7 @@
 package com.sanaa.tudee_assistant.presentation.screen.main
 
 import com.sanaa.tudee_assistant.domain.service.PreferencesManager
-import com.sanaa.tudee_assistant.presentation.utils.BaseViewModel
+import com.sanaa.tudee_assistant.presentation.base.BaseViewModel
 import kotlinx.coroutines.flow.update
